Move explain_queries to indexdigest.utils
