+++ indexdigest/linters/not_used_indices.py
@@ -5,24 +5,7 @@ import logging
 
 from collections import defaultdict, OrderedDict
 
-from indexdigest.utils import LinterEntry, is_select_query
-
-
-def explain_queries(database, queries):
-    """
-    Yields EXPLAIN result rows for given queries
-
-    :type database  indexdigest.database.Database
-    :type queries list[str]
-    :rtype: tuple[str,str,str,str]
-    """
-    # analyze only SELECT queries from the log
-    for query in filter(is_select_query, queries):
-        for row in database.explain_query(query):
-            table_used = row['table']
-            index_used = row['key']
-
-            yield (query, table_used, index_used, row)
+from indexdigest.utils import LinterEntry, explain_queries
 
 
 def check_not_used_indices(database, queries):
